% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustedPAS.R
\name{adjustedPAS}
\alias{adjustedPAS}
\title{Adjusted PAS}
\usage{
adjustedPAS(Query_Time, Reference_Aligned_Time_Median, PAS_Abs)
}
\arguments{
\item{Query_Time}{The query time (vector)}

\item{Reference_Aligned_Time_Median}{The median aligned reference time (after truncation) of all STP genes (grouped by each query time).}

\item{PAS_Abs}{Absolute PAS value}
}
\value{
This function returns Adjusted PAS
}
\description{
This function calcuates the adjusted PAS. The adjustedPAS value is to adjust systematic difference between two conditions, such as species.
}
\examples{
data(similarHumanMouse)
alignableRegion_timePoints_query_merged_all=c()
alignableRegion_timePoints_reference_merged_all=c()
for (k in names(similarHumanMouse)) {
    alignableRegion_timePoints_query_merged_all=c(alignableRegion_timePoints_query_merged_all,similarHumanMouse[[k]][["alignableRegion_timePoints_query_merged"]])
    alignableRegion_timePoints_reference_merged_all=c(alignableRegion_timePoints_reference_merged_all,similarHumanMouse[[k]][["alignableRegion_timePoints_reference_merged"]])
}
Query_to_median_Reference=tapply(alignableRegion_timePoints_reference_merged_all,as.factor(alignableRegion_timePoints_query_merged_all),median)
Query_Time=as.numeric(names(Query_to_median_Reference))
Reference_Aligned_Time_Median=as.numeric(Query_to_median_Reference)

PAS=similarHumanMouse[['ASIC4']]$PAS
PAS_adjusted=adjustedPAS(Query_Time,Reference_Aligned_Time_Median,PAS)
PAS
PAS_adjusted

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
