% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignableRegionIndex.R
\name{alignableRegionIndex}
\alias{alignableRegionIndex}
\title{Obtain alignable region indices}
\usage{
alignableRegionIndex(aligned_timePoints_query,
  aligned_timePoints_reference)
}
\arguments{
\item{aligned_timePoints_query}{A vector containing DTW aligned Time Points (query)}

\item{aligned_timePoints_reference}{A vector containing DTW aligned Time Points (reference)}
}
\value{
Indices of alignable region (after truncation)
}
\description{
This function corrects the DTW aligned indices by masking the first (m-1) start time points in one gene if the first m time points can be aligned to the same start points in another gene, and terminating alignment if DTW matches the last elements in any of the genes.
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference

alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]

index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)

alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
