% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAUC.R
\name{getAUC}
\alias{getAUC}
\title{Area under the curve for a linear regression}
\usage{
getAUC(x_start, x_end, slope, intercept)
}
\arguments{
\item{x_start}{start coordinate for a linear regression}

\item{x_end}{end coordinate for a linear regression}

\item{slope}{slope of a linear regression}

\item{intercept}{intercept of a linear regression}
}
\value{
area under the curve for a linear regression
}
\description{
This function calculate the area under the curve for a linear regression
}
\examples{
x_start=0
x_end=10
slope=1
intercept=1
getAUC(x_start, x_end, slope, intercept)

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
