% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPValueRho.R
\name{getPValueRho}
\alias{getPValueRho}
\title{Likelihood of alignment arising by chance}
\usage{
getPValueRho(Rho, query, timePoints_query, reference, timePoints_reference)
}
\arguments{
\item{Rho}{Aligned gene expression correlation by calculating the Spearman's rank correlation coefficient (Rho) for reliable aligned gene expressions (after truncation)}

\item{query}{A vector of timecourse gene expression values of query}

\item{timePoints_query}{A vector of time points of query}

\item{reference}{A vector of timecourse gene expression values of reference}

\item{timePoints_reference}{A vector of time points of reference}
}
\value{
p-value of likelihood of alignment arising by chance
}
\description{
This function calculate the likelihood of temporal alignment arising by chance.
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference
alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]
index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)
alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]
percentageAlignmentQuery=percentageAlignment(timePoints_query,
                         timePoints_reference,
                         alignableRegion_timePoints_query,
                         alignableRegion_timePoints_reference)['percentage_alignment_query']
percentageAlignmentReference=percentageAlignment(timePoints_query,
                             timePoints_reference,
                             alignableRegion_timePoints_query,
                             alignableRegion_timePoints_reference)['percentage_alignment_reference']
Rho=spearmanCorrelation(alignableRegion_values_query,alignableRegion_values_reference)
pValueRho=getPValueRho(Rho,query,timePoints_query,reference,timePoints_reference)

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
