% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeReferencePoints.R
\name{mergeReferencePoints}
\alias{mergeReferencePoints}
\title{Calculating the average of aligned time of reference for each aligned query time point}
\usage{
mergeReferencePoints(alignableRegion_timePoints_query,
  alignableRegion_timePoints_reference)
}
\arguments{
\item{alignableRegion_timePoints_query}{A vector: alignable time points (after trunction) of query}

\item{alignableRegion_timePoints_reference}{A vector: alignable time points (after trunction) of reference}
}
\value{
\item{aligned_timePoints_query_merged }{alignable time points (after trunction) of query, the same as input}
\item{aligned_timePoints_reference_merged }{average aligned time (after trunction) of reference for each aligned (after trunction) query}
}
\description{
The function groups the reference aligned time points based on query aligned time points. Then it calculate the average aligned time (reference) for each query time point (aligned).
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference
alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]
index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)
alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]
percentageAlignmentQuery=percentageAlignment(timePoints_query,
               timePoints_reference,
               alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)['percentage_alignment_query']
percentageAlignmentReference=percentageAlignment(timePoints_query,
               timePoints_reference,
               alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)['percentage_alignment_reference']
Rho=spearmanCorrelation(alignableRegion_values_query,alignableRegion_values_reference)
pValueRho=getPValueRho(Rho,query,timePoints_query,reference,timePoints_reference)
alignableRegion_timePoints_query_merged=mergeReferencePoints(alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)["aligned_timePoints_query_merged"][[1]]
alignableRegion_timePoints_reference_merged=mergeReferencePoints(alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)["aligned_timePoints_reference_merged"][[1]]

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
