% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentageAlignment.R
\name{percentageAlignment}
\alias{percentageAlignment}
\title{Percentage of Alignment}
\usage{
percentageAlignment(timePoints_query, timePoints_reference,
  alignableRegion_timePoints_query, alignableRegion_timePoints_reference)
}
\arguments{
\item{timePoints_query}{A vector containing time points (Query)}

\item{timePoints_reference}{A vector containing time points (Reference)}

\item{alignableRegion_timePoints_query}{A vector containing truncated time points (Query)}

\item{alignableRegion_timePoints_reference}{A vector containing truncated time points (Reference)}
}
\value{
\item{percentage_alignment_query }{percentage of alignment for query}
 \item{percentage_alignment_reference }{percentage of alignment for reference}
}
\description{
This funciton calcualtes the percentage of alignment for query and for reference, respectively: the length of reliable aligned time interval in query (or in reference) divided by total length of query (or reference) time interval.
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference
alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]
index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)
alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]
percentageAlignmentQuery=percentageAlignment(timePoints_query,
                         timePoints_reference,
                         alignableRegion_timePoints_query,
                         alignableRegion_timePoints_reference)['percentage_alignment_query']
percentageAlignmentReference=percentageAlignment(timePoints_query,
                             timePoints_reference,
                             alignableRegion_timePoints_query,
                             alignableRegion_timePoints_reference)['percentage_alignment_reference']

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
