% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPairwiseTimePoints.R
\name{plotPairwiseTimePoints}
\alias{plotPairwiseTimePoints}
\title{Plot aligned time points}
\usage{
plotPairwiseTimePoints(alignableRegion_timePoints_query_merged,
  alignableRegion_timePoints_reference_merged, breakPointsMatrix,
  title = "Aligned Time Points", dot_pch = 1, dot_cex = 0.6,
  line_lwd = 1, abline_lty = 3, abline_lwd = 2,
  xlabText = "Time (Query)", ylabText = "Time (Reference)",
  cex_lab = 1)
}
\arguments{
\item{alignableRegion_timePoints_query_merged}{Output from mergeReferencePoints()["aligned_timePoints_query_merged"][[1]]}

\item{alignableRegion_timePoints_reference_merged}{Output from mergeReferencePoints()["aligned_timePoints_reference_merged"][[1]]}

\item{breakPointsMatrix}{Output from fetchBreakPoints()}

\item{title}{Title of the figure}

\item{dot_pch}{Plot symbols (R standard pch parameter). 1=default.}

\item{dot_cex}{Magnification of plot symbols. 0.6=default.}

\item{line_lwd}{Plot line width. 1=default.}

\item{abline_lty}{Diagonal line line type. 1:solid, 2:dashed, 3:dotted, 4:dotdash, 5:longdash, 6:twodash. 3=default}

\item{abline_lwd}{Diagonal line line width. 2=default.}

\item{xlabText}{X-axis label text. "Time (Query)"=default}

\item{ylabText}{Y-axis label text. "Time (Reference)"=default}

\item{cex_lab}{Magnification of x and y labels relative to cex. 1=default}
}
\value{
It returns a plot.
}
\description{
This funciton plots aligned time points (after trunction)
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference
alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]
index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)
alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]
percentageAlignmentQuery=percentageAlignment(timePoints_query,timePoints_reference,
               alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)['percentage_alignment_query']
percentageAlignmentReference=percentageAlignment(timePoints_query,
               timePoints_reference,
               alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)['percentage_alignment_reference']
Rho=spearmanCorrelation(alignableRegion_values_query,alignableRegion_values_reference)
pValueRho=getPValueRho(Rho,query,timePoints_query,reference,timePoints_reference)
alignableRegion_timePoints_query_merged=mergeReferencePoints(alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)["aligned_timePoints_query_merged"][[1]]
alignableRegion_timePoints_reference_merged=mergeReferencePoints(alignableRegion_timePoints_query,
               alignableRegion_timePoints_reference)["aligned_timePoints_reference_merged"][[1]]
segmentedRegression_out=segmentedRegression(alignableRegion_timePoints_query_merged,
               alignableRegion_timePoints_reference_merged)
breakPointsMatrix=fetchBreakPoints(segmentedRegression_out)
plotPairwiseTimePoints(alignableRegion_timePoints_query_merged,
               alignableRegion_timePoints_reference_merged,
               breakPointsMatrix,
               title=gene)

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
