% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentedRegression.R
\name{segmentedRegression}
\alias{segmentedRegression}
\title{Segmented Regression For Aligned Time Points}
\usage{
segmentedRegression(aligned_timePoints_query, aligned_timePoints_reference,
  timeDigits = 2, deltaSlope = 0.1, N = 10)
}
\arguments{
\item{aligned_timePoints_query}{alignable time points (after trunction) of query}

\item{aligned_timePoints_reference}{alignable time points (after trunction) of reference}

\item{timeDigits}{Precision (2=default)}

\item{deltaSlope}{Merge two segments into one of splop difference less than deltaSlope (0.1=default).}
}
\value{
A list containing segments informatioin. Each segment is a list containing ("x_start","x_end","slope","intercept") information.
 \item{x_start}{break point start}
 \item{x_end}{break point end}
 \item{slop}{slop of regression}
 \item{intercept}{intercept of of regression}
}
\description{
This function partitions aligned time points into different segments
}
\examples{
data(simData)
data=simdata$TimeShift_10
gene=data$gene
query=data$query
timePoints_query=data$timePoints_query
reference=data$reference
timePoints_reference=data$timePoints_reference
alignment=dtw(query,reference)
dtw_results=list(alignment$index1,alignment$index2)
index_1=dtw_results[[1]]
index_2=dtw_results[[2]]
aligned_values_query=query[index_1]
aligned_values_reference=reference[index_2]
aligned_timePoints_query=timePoints_query[index_1]
aligned_timePoints_reference=timePoints_reference[index_2]
index_alignableRegion=alignableRegionIndex(aligned_timePoints_query,aligned_timePoints_reference)
alignableRegion_values_query=aligned_values_query[index_alignableRegion]
alignableRegion_values_reference=aligned_values_reference[index_alignableRegion]
alignableRegion_timePoints_query=aligned_timePoints_query[index_alignableRegion]
alignableRegion_timePoints_reference=aligned_timePoints_reference[index_alignableRegion]
percentageAlignmentQuery=percentageAlignment(timePoints_query,
         timePoints_reference,
         alignableRegion_timePoints_query,
         alignableRegion_timePoints_reference)['percentage_alignment_query']
percentageAlignmentReference=percentageAlignment(timePoints_query,
         timePoints_reference,
         alignableRegion_timePoints_query,
         alignableRegion_timePoints_reference)['percentage_alignment_reference']
Rho=spearmanCorrelation(alignableRegion_values_query,alignableRegion_values_reference)
pValueRho=getPValueRho(Rho,query,timePoints_query,reference,timePoints_reference)
alignableRegion_timePoints_query_merged=mergeReferencePoints(alignableRegion_timePoints_query,
         alignableRegion_timePoints_reference)["aligned_timePoints_query_merged"][[1]]
alignableRegion_timePoints_reference_merged=mergeReferencePoints(alignableRegion_timePoints_query,
         alignableRegion_timePoints_reference)["aligned_timePoints_reference_merged"][[1]]
segmentedRegression_out=segmentedRegression(alignableRegion_timePoints_query_merged,
         alignableRegion_timePoints_reference_merged)

}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
