% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated Timecourse datasets}
\usage{
data(simData)
}
\description{
Containing 4 datasets: "TimeShift_10", "TimeShift_20", "Different_speed_1.5", and "Complex".
}
\details{
\itemize{
  \item gene: Gene name
  \item query: Timecourse Gene expression values of query
  \item timePoints_query: Time Points of of query
  \item reference: Timecourse Gene expression values of reference
  \item timePoints_reference: Time Points of of reference
}
}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
\keyword{datasets}
\keyword{simulated}
