% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarHumanMouse.R
\docType{data}
\name{similarHumanMouse}
\alias{similarHumanMouse}
\title{Human-Mouse Timecourse datasets (Temporal Similar Genes)}
\usage{
data(similarHumanMouse)
}
\description{
Human-Mouse Timecourse datasets (Temporal Similar Genes)
}
\details{
\itemize{
  \item gene: Gene name
  \item alignableRegion_timePoints_query_merged: aligned query time 
  \item alignableRegion_timePoints_reference_merged: aligned reference time 
  \item PAS: absolute PAS value
  
}
}
\author{
Peng Jiang \email{PJiang@morgridge.org}
}
\keyword{Human-Mouse}
\keyword{datasets}
